refactor: use jest configurator package, so refactored the debounce tests
