+++ packages/provider/__tests__/inner.js
@@ -1,0 +1,24 @@
+import { Component } from "react";
+import PropTypes from "prop-types";
+
+class Inner extends Component {
+  constructor(props) {
+    super(props);
+    this.numberOfDebouncedPropsUpdates = 0;
+  }
+
+  componentWillReceiveProps({ debouncedProps }) {
+    if (debouncedProps !== this.props.debouncedProps) {
+      this.numberOfDebouncedPropsUpdates += 1;
+    }
+  }
+
+  render() {
+    return "hello";
+  }
+}
+Inner.propTypes = {
+  debouncedProps: PropTypes.shape({}).isRequired
+};
+
+export default Inner;
